User- and Tag policies

+++ app/Http/Controllers/TagController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Tag;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TagController extends Controller
 {
@@ -72,6 +73,9 @@ class TagController extends Controller
      */
     public function edit(Tag $tag)
     {
+
+        abort_unless(Gate::allows('update', $tag), 403);
+
         return view('tag.edit')->with([
             'tag' => $tag
         ]);
@@ -86,6 +90,8 @@ class TagController extends Controller
      */
     public function update(Request $request, Tag $tag)
     {
+        abort_unless(Gate::allows('update', $tag), 403);
+
         $request->validate([
             'name' => 'required',
             'style' => 'required',
@@ -111,6 +117,9 @@ class TagController extends Controller
      */
     public function destroy(Tag $tag)
     {
+
+        abort_unless(Gate::allows('delete', $tag), 403);
+
         $oldName = $tag->name;
         $tag->delete();
         return $this->index()->with(
